Bump integer size

+++ sql/package_stats.sql
@@ -5,7 +5,7 @@
 DROP TABLE IF EXISTS `package_stats`;
 
 CREATE TABLE `package_stats` (
-  `dl_number` mediumint(8) unsigned NOT NULL default '0',
+  `dl_number` int(11) unsigned NOT NULL default '0',
   `package` varchar(80) NOT NULL default '',
   `release` varchar(20) NOT NULL default '',
   `pid` int(11) NOT NULL default '0',
